+++ base_exts/test/fn.ml
@@ -78,3 +78,9 @@ let%expect_test "never example" =
 let%expect_test "never example (different type, to show polymorphism)" =
   printf "%b\n" (never "say never again") ;
   [%expect {| false |}]
+
+let%expect_test "Compose_syntax: >> example" =
+  let f : string -> bool = String.(Compose_syntax.(strip >> lowercase >> equal "test")) in
+  printf "%b\n" (f "  TEST  ") ;
+  [%expect {| true |}]
+